// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] }
// ...
export interface BlogPostDocumentDataTagsItem {
  /**
   * Tag field in *Blog Post → Tags*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Code
   * - **API ID Path**: blog_post.tags[].tag
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  tag: prismic.KeyTextField
}

/**
 * Content for Blog Post documents
 */
interface BlogPostDocumentData {
  /**
   * Title field in *Blog Post*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Guardian of the Digital Realm: Web Security
   * - **API ID Path**: blog_post.title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  title: prismic.RichTextField

  /**
   * Excerpt field in *Blog Post*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: A brief summary of the blog post...
   * - **API ID Path**: blog_post.excerpt
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  excerpt: prismic.RichTextField

  /**
   * Featured Image field in *Blog Post*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.featured_image
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  featured_image: prismic.ImageField<never>

  /**
   * Author Name field in *Blog Post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Jese Leos
   * - **API ID Path**: blog_post.author_name
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  author_name: prismic.KeyTextField

  /**
   * Author Avatar field in *Blog Post*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.author_avatar
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  author_avatar: prismic.ImageField<never>

  /**
   * Publish Date field in *Blog Post*
   *
   * - **Field Type**: Date
   * - **Placeholder**: Select publish date
   * - **API ID Path**: blog_post.publish_date
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/date
   */
  publish_date: prismic.DateField

  /**
   * Badge Type field in *Blog Post*
   *
   * - **Field Type**: Select
   * - **Placeholder**: *None*
   * - **Default Value**: Article
   * - **API ID Path**: blog_post.badge_type
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/select
   */
  badge_type: prismic.SelectField<'Tutorial' | 'Article', 'filled'>

  /**
   * Tags field in *Blog Post*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.tags[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/repeatable-group
   */
  tags: prismic.GroupField<Simplify<BlogPostDocumentDataTagsItem>>

  /**
   * Content field in *Blog Post*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Write your blog post content here...
   * - **API ID Path**: blog_post.content
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  content: prismic.RichTextField
}

/**
 * Blog Post document from Prismic
 *
 * - **API ID**: `blog_post`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type BlogPostDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<BlogPostDocumentData>,
    'blog_post',
    Lang
  >